# lib/api-zod/src/generated/types/coachContext.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Qoot nutrition and fitness coach API
 * OpenAPI spec version: 0.1.0
 */
import type { CoachTurn } from './coachTurn';
import type { Dashboard } from './dashboard';
import type { Plan } from './plan';
import type { Profile } from './profile';

export interface CoachContext {
  profile?: Profile;
  plan?: Plan;
  dashboard?: Dashboard;
  recentMessages?: CoachTurn[];
}